test and cleanup

+++ artemis/src/main/java/tech/pegasys/artemis/state/util/AttestationUtil.java
@@ -17,6 +17,7 @@ import tech.pegasys.artemis.Constants;
 import tech.pegasys.artemis.datastructures.beaconchainstate.PendingAttestationRecord;
 import tech.pegasys.artemis.ethereum.core.Hash;
 import tech.pegasys.artemis.state.BeaconState;
+import tech.pegasys.artemis.util.uint.UInt64;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -61,7 +62,7 @@ public class AttestationUtil {
     public static Hash get_block_root(BeaconState state, long slot) throws Exception {
         long slot_upper_bound = slot + state.getLatest_block_roots().size();
         if((state.getSlot() <= slot_upper_bound) || slot < state.getSlot())
-            return state.getLatest_block_roots().get(Math.toIntExact(slot % state.getLatest_block_roots().size()));
+            return state.getLatest_block_roots().get(UInt64.valueOf(slot % state.getLatest_block_roots().size()));
         throw new BlockValidationException("Desired block root not within the provided bounds");
     }
